[DATA-refactor] Use different waitgroup for capture(). (#819)
Previously the service's overall WaitGroup was used. This is unnecessary: since these goroutines are being spun off by a single goroutine (the c.capture() goroutine - which itself is a background worker of c), their waitgroup can be confined to the scope of those capture methods (sleep and tickerBasedCapture). This nesting of WaitGroups vs using a shared WaitGroup reduces the risk of race conditions and allowed me to remove the use of c.lock when adding to the waitgroup within the captyre methods.