[Data-166]: Implement and Test Client for File & Sensor Uploads (#915)
Client for File & Sensor Uploads. Now supports arbitrary file type uploads and chunking of that data (when the data is very large it is better to upload in smaller chunks - this will be helpful for future TODOs like partial file uploading). Includes testing for binary & tabular sensor data uploads as well.