[RSDK 402] add default starting position attribute to servos (#981)
Adds a starting position attribute that a user can set depending on where the servo is positioned at startup.

A hold position bool is added, that defaults true for holding the servo position. If it is false, the servo holds its position for 500ms. This protects a servo from burning out if it hits an obstacle.

Also addresses RDK 434, the opManager for Servo is refined.