[RSDK-2411] update LastError to allow suppression of rare errors (#2129)
Co-authored-by: randhid <35934754+randhid@users.noreply.github.com>

The `LastError` struct used to store the most recent error, and return and clear it when asked. This commit modifies `LastError` to take 2 parameters: we'll only return the most recent (non-nil) error if at least M of the last N errors stored inside us was non-nil. You can get the original behavior by setting both values to 1, which is what I've done in most of the places it was used. The movement sensors that communicate over I2C have been modified so that they only give back errors if at least 5 of the last 10 values were non-nil, so that a flaky I2C bus doesn't give spurious trouble. Are those good values to have? I dunno; I still can't reproduce the flaky behavior.

I've also added a test, intending to show what the behavior is supposed to be. The new test (and all the old tests) passes, and the ADXL and MPU sensors still work fine when they're healthy (I'm still unable to reproduce the flakiness that prompted this).